refactor: modularizar dominio tables em api/model/hooks/ui
