+++ src/domains/tables/hooks/tables-query-keys.ts
@@ -1,0 +1,5 @@
+export const tablesQueryKeys = {
+	all: ["tables"] as const,
+	list: (organizationId: string | null) =>
+		[...tablesQueryKeys.all, "list", organizationId] as const,
+}
